LPS-199532 Extract service from NPMJavaScriptAwarePortalWebResources to NPMPortalWebResources and extends NPMPortalWebResources in NPMJavaScriptAwarePortalWebResources
