+++ modules/apps/frontend-js/frontend-js-web/bnd.bnd
@@ -17,4 +17,5 @@ Provide-Capability:\
 		version:Version="${Bundle-Version}"
 Web-ContextPath: /frontend-js-web
 -donotcopy: (\.touch|.*_diffs.*)
+-dsannotations-options: inherit
 -includeresource: META-INF/resources=tmp/META-INF/resources
