+++ modules/apps/frontend-js/frontend-js-web/src/main/java/com/liferay/frontend/js/web/internal/modules/extender/npm/NPMPortalWebResources.java
@@ -1,0 +1,59 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2023 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+package com.liferay.frontend.js.web.internal.modules.extender.npm;
+
+import com.liferay.portal.kernel.servlet.PortalWebResourceConstants;
+import com.liferay.portal.kernel.servlet.PortalWebResources;
+
+import java.util.concurrent.atomic.AtomicLong;
+
+import javax.servlet.ServletContext;
+
+import org.osgi.framework.Bundle;
+import org.osgi.framework.BundleContext;
+import org.osgi.service.component.annotations.Activate;
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * @author Joao Victor Alves
+ */
+@Component(service = PortalWebResources.class)
+public class NPMPortalWebResources implements PortalWebResources {
+
+	@Override
+	public String getContextPath() {
+		return servletContext.getContextPath();
+	}
+
+	@Override
+	public long getLastModified() {
+		return lastModified.get();
+	}
+
+	@Override
+	public String getResourceType() {
+		return PortalWebResourceConstants.RESOURCE_TYPE_JS;
+	}
+
+	@Override
+	public ServletContext getServletContext() {
+		return servletContext;
+	}
+
+	@Activate
+	protected void activate(BundleContext bundleContext) {
+		Bundle bundle = bundleContext.getBundle();
+
+		lastModified.set(bundle.getLastModified());
+	}
+
+	protected final AtomicLong lastModified = new AtomicLong();
+
+	@Reference(target = "(osgi.web.symbolicname=com.liferay.frontend.js.web)")
+	protected ServletContext servletContext;
+
+}
